+++ sphinx/source/changelog.rst
@@ -5,6 +5,8 @@ Changelog
 -----
 * Added the ``--search`` option to :ref:`skool2html.py` (to add a directory to
   the resource search path)
+* Added the ``--writer`` option to :ref:`skool2html.py` (for specifying the
+  HTML writer class to use)
 * Added the ``LinkInternalOperands`` parameter to the :ref:`ref-Game` section
   (for specifying whether to hyperlink instruction operands that refer to an
   address in the same entry)